// src/lib/api-client/models/rejection-summary-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * DomainFlow is a campaign-focused domain intelligence platform. The API manages personas, proxies, keyword pipelines, monitoring, and campaign execution while exposing configuration and scoring surfaces for both UI and automation clients. 
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: api@domainflow.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RejectionSummaryResponseCounts } from './rejection-summary-response-counts';
// May contain unused imports in some cases
// @ts-ignore
import type { RejectionSummaryResponseTotals } from './rejection-summary-response-totals';

/**
 * Breakdown of domain outcomes by rejection_reason. Enables audit equation: analyzed = qualified + rejected_total (low_score + no_keywords + parked + dns errors + http errors) 
 * @export
 * @interface RejectionSummaryResponse
 */
export interface RejectionSummaryResponse {
  /**
   * 
   * @memberof RejectionSummaryResponse
   */
  'campaignId': 
        
          
          string
    ;
  /**
   * 
   * @memberof RejectionSummaryResponse
   */
  'counts': 
        
          
          RejectionSummaryResponseCounts
    ;
  /**
   * 
   * @memberof RejectionSummaryResponse
   */
  'totals': 
        
          
          RejectionSummaryResponseTotals
    ;
  /**
   * True if audit equation balances: analyzed == qualified + rejected. False indicates data inconsistency requiring investigation. 
   * @memberof RejectionSummaryResponse
   */
  'balanced': 
        
          
          boolean
    ;
  /**
   * Human-readable explanation if balanced is false
   * @memberof RejectionSummaryResponse
   */
  'auditNote'?: 
        
          
          string
    ;
}