// src/lib/api-client/models/score-component.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * DomainFlow is a campaign-focused domain intelligence platform. The API manages personas, proxies, keyword pipelines, monitoring, and campaign execution while exposing configuration and scoring surfaces for both UI and automation clients. 
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: api@domainflow.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Individual component score with state and optional reason
 * @export
 * @interface ScoreComponent
 */
export interface ScoreComponent {
  /**
   * Normalized score (0-1), null if unavailable
   * @memberof ScoreComponent
   */
  'value'?: 
        
          
          number
    ;
  /**
   * Component availability state
   * @memberof ScoreComponent
   */
  'state': 
      'ok' | 'unavailable' | 'error'
;
  /**
   * Reason if state is not ok
   * @memberof ScoreComponent
   */
  'reason'?: 
      'field_missing' | 'computation_failed' | 'data_pending'
;
}